elapsed time

+++ msgf/entry.py
@@ -2,6 +2,7 @@
 # EXTRACT, IF NECESSARY, AND RUN MGSF PLUS TO 
 # NORMALIZE THE DATA
 # DANIEL.KRISTIYANTO@PNNL.GOV
+# Docker container: msgf
 
 ######################## ENV ###########################
 import gzip
@@ -21,6 +22,7 @@ out 		= []
 db 			= None
 module 		= None
 input_csv	= None
+start_time = time.time()
 
 ######################## DECOMPRESS ###########################
 spectrum_tmp = []
@@ -28,11 +30,12 @@ for src_name in glob.glob(os.path.join(working_dir, '*.gz')):
     base = os.path.basename(src_name)
     print("Extracting", src_name)
     dest_name = os.path.join(working_dir, base[:-3])
-    spectrum_tmp.append(dest_name)
-    with gzip.open(src_name, 'rb') as infile:
-        with open(dest_name, 'wb') as outfile:
-            for line in infile:
-                outfile.write(line)
+    if not os.path.isfile(dest_name):
+    	spectrum_tmp.append(dest_name)
+    	with gzip.open(src_name, 'rb') as infile:
+        	with open(dest_name, 'wb') as outfile:
+            	for line in infile:
+                	outfile.write(line)
 
 spectrum 				= scan_spectrum(working_dir)
 db, input_csv 			= scan_dir(working_dir)
@@ -56,6 +59,10 @@ db, input_csv 			= scan_dir(working_dir)
 for s in spectrum:
 	out = os.path.splitext(s)[0]+'.mzid'
 	msgf(s,db,out)
+	print("Done.")
+
+elapsed_time = time.time() - start_time
+print("Elapsed time:", str(timedelta(seconds=elapsed_time)))
 
 # if len(scan_mzid(working_dir)) != 0:
 # 	try:
